projects/auto_affiliate_video/requirements.txt
moviepy
edge-tts
openai
python-dotenv
pillow
